Location and Group classes
Added Location and Group classes to store data in them. Added Locations
table to the database and functions for adding a new location+tying it
to an user. Added functions to retrieve Locations from the table by
location id, user id and the latest location of an user (needs to be
tested). Removed Lat and Lng fields from the User class. Added DBBrowser
activity that lists out all user- and location information associated
with that user.